+++ CHANGELOG.md
@@ -1,5 +1,9 @@
  Luminati Proxy manager - Change Log
 
+## 1.81.930
+- :sparkles: improved configuration tab and added download option
+- :sparkles: UI and bug fixing
+
 ## 1.81.318
 - :star: add min request time as rule trigger type
 - :sparkles: improved multiply feature